remove unused es dir, create new esClient module
