+++ modules/server/src/server.ts
@@ -1,4 +1,3 @@
-import { Client, type ClientOptions } from '@elastic/elasticsearch';
 import { Router, type Request, type Response, type NextFunction } from 'express';
 import morgan from 'morgan';
 
@@ -19,32 +18,6 @@ const {
 	ES_LOG, //TODO: ES doesn't include a logger anymore
 	PING_PATH,
 } = ENV_CONFIG;
-
-export const buildEsClient = (esHost = '', esUser = '', esPass = '') => {
-	if (!esHost) {
-		console.error('no elasticsearch host was provided');
-	}
-
-	const esConfig: ClientOptions = {
-		node: esHost,
-	};
-
-	if (esUser) {
-		if (!esPass) {
-			console.error('ES user was defined, but password was not');
-		}
-		esConfig['auth'] = {
-			username: esUser,
-			password: esPass,
-		};
-	}
-
-	return new Client(esConfig);
-};
-
-export const buildEsClientViaEnv = () => {
-	return buildEsClient(ES_HOST, ES_USER, ES_PASS);
-};
 
 const arrangerServer = async ({
 	configsSource = CONFIG_FILES_PATH,
